Validate CSV templates at input startup time
It is better to fail fast (and display the failure right after saving the input) than is it to delay failure until message processing.

The Graylog input system does not have custom validation support on the Edit Input dialog. The next best thing is to throw a MisfireException on input startup (happens right after save anyhow). This way, the user will see that something is wrong with the config right away.