Add Kinesis Healthcheck (#45)
* Improve organization for Flow Log message detection

* Improve Flow Log test

It now tests for a message with too many and too few spaces.

* Add TODOs for healthCheck method logic

* Add beginnings of Kinesis healthChecker

This will pull a establish a Kinesis subscription and pull a single message from a Kinesis stream.

* Fix failing unit test

* Continue developing KinesisHealthCheck

- Remove unneeded metric tracking
- Remove extra parsing logic (this object should just hand back payload and not do any parsing)
- Improve application name handling
- Add comments

* Add detection logic for raw vs. CloudWatch logs

* Remove KinesisHealthCheck class

The KinesisConsumer does not work well for the health check (designed for realtime processing, takes a long time to start, cannot detect empty stream, and is really hard to use in a quick API request). Now, we're planning to directly retrieve the messages using the Kinesis client. This is the most straight-forward thing  We might revisit this later.

* Fix JSON parsing of Kinesis CloudWatch subscription record

Parse the record just as was done in the existing AWS plugin. The logic now includes autodetection of compressed/vs not compressed. Mock Kinesis CloudWatch subscription record included for testing purposes.

* Add CloudWatch logs codec and tests from existing AWS plugin

* Parse Flow Log message into object

* Load appropriate codec during healthCheck process

When the message type is detected, load the respective codec for that message type.

* Parse message with appropriate codec

Once the log message type is detected, then the codec is looked up. The message is then parsed with the codec.

* Supply log group name with the response

* Improve comments, logging, and error checking

The log group name is now also included in the response.

* Add Flow Log codec test

* Use AutoValue for CloudWatchLogEntry class

* Use AutoValue for all remaining CloudWatch value classes

* Cleanup merge conflicts after rebasing

* Specify constants for all JsonProperty annotations

* Delete uneeded KinesisDTO

All data will be stored in the input

* Establish a base AWSRequest JSON class

* Fix Guice injection error for KinesisService

* Add sample cURL command for healthCheck method

A similar cURL command will be used for other methods, so that it is clear how the UI will use them.

* Remove unneeded Kinesis Client 1.x dependency

* Add formatted message summary in the Health Check response

* Cleanup formatting and TODOs