AWS Cloudwatch backend API calls
Add initial backend API calls for CloudWatch integration: getRegions, getLogGroups, getStreams, retrieveKinesisMessages, healthCheck

* AWS Cloud Watch services and resources (#24)

Adds beginnings of API endpoint and Kinesis/Cloudwatch services

Includes a structure that we will continue to build from.

* Rework organization of classes for unified structure

The goal is to establish some structure that we can implement AWS API calls within. There's now one resource (for API calls), one service (for business logic), and one AWSClient (for all AWS/API SDK interactions).

* Cleanup and add comments

* Cleanup code around log message auto-detection

* Aws cloudwatch client (#31)

* Add CloudWatchService class

* Add AWSConfigSettings class

* Add UserCredentials class

* Add temporary Main class

* Include the latest version of the Apache Http Client

* Improve HTTP Client dependency comment

* Update AWSConfigSettings class

* Update CloudWatchService class

* Update temporary Main class

* Fix inject error in AWSResource

* Fix merge conflicts

* Restructure Resources and Service classes

* Add getRegion into AWSResource (#43)

* Add getRegion into AWSResource

* Update api paths

* Increase max retry limit for stream get to 1000

Runaway looping Stopping at 100 is probably to small.

* Use underscores instead of camel case for json names

It is a general project standard to use underscores.

* Add paging functionality to getLogGroupNames (#53)

* Add paging functionality to getLogGroupNames

* Add unit test for CloudWatch log groups

* Add comments in log group name unit test

* Code clean up and remove unneeded code

* Revert unintended change from unit test commit

* Add retrievelogs for Kinesis Healthcheck (#76)

* Refactor getKinesisStreams in KinesisService

* Add validCredentials method in AWSService class

* Temporary Main class added to test putting records into a Kinesis stream

* Add retrieveKinesisLogs in KinesisService

* Update temporary Main class

* Update retrieveKinesisLogs to loop through shard iterators

* Update Main class

* Update pom file

* Update validateCredentials in AWSService class

* Add createKinesisClient method in KinesisService class

* Add testGetStreamCredentials and update testGetStreams

* Healthcheck merge (#81)

* empty commit to push branch

* Add Kinesis Healthcheck (#45)

* Improve organization for Flow Log message detection

* Improve Flow Log test

It now tests for a message with too many and too few spaces.

* Add TODOs for healthCheck method logic

* Add beginnings of Kinesis healthChecker

This will pull a establish a Kinesis subscription and pull a single message from a Kinesis stream.

* Fix failing unit test

* Continue developing KinesisHealthCheck

- Remove unneeded metric tracking
- Remove extra parsing logic (this object should just hand back payload and not do any parsing)
- Improve application name handling
- Add comments

* Add detection logic for raw vs. CloudWatch logs

* Remove KinesisHealthCheck class

The KinesisConsumer does not work well for the health check (designed for realtime processing, takes a long time to start, cannot detect empty stream, and is really hard to use in a quick API request). Now, we're planning to directly retrieve the messages using the Kinesis client. This is the most straight-forward thing  We might revisit this later.

* Fix JSON parsing of Kinesis CloudWatch subscription record

Parse the record just as was done in the existing AWS plugin. The logic now includes autodetection of compressed/vs not compressed. Mock Kinesis CloudWatch subscription record included for testing purposes.

* Add CloudWatch logs codec and tests from existing AWS plugin

* Parse Flow Log message into object

* Load appropriate codec during healthCheck process

When the message type is detected, load the respective codec for that message type.

* Parse message with appropriate codec

Once the log message type is detected, then the codec is looked up. The message is then parsed with the codec.

* Supply log group name with the response

* Improve comments, logging, and error checking

The log group name is now also included in the response.

* Add Flow Log codec test

* Use AutoValue for CloudWatchLogEntry class

* Use AutoValue for all remaining CloudWatch value classes

* Cleanup merge conflicts after rebasing

* Specify constants for all JsonProperty annotations

* Delete uneeded KinesisDTO

All data will be stored in the input

* Establish a base AWSRequest JSON class

* Fix Guice injection error for KinesisService

* Add sample cURL command for healthCheck method

A similar cURL command will be used for other methods, so that it is clear how the UI will use them.

* Remove unneeded Kinesis Client 1.x dependency

* Add formatted message summary in the Health Check response

* Cleanup formatting and TODOs

* Minor cleanup after rebasing and merging

* Fix failing unit tests