Add missing permissions for sample AWS policy (#122)
* Migrate AWS policy to autovalue object

This avoids embedding JSON in a Java constant, which is error prone.

* Add missing permissions

Added:
- kinesis:ListStreams
- logs:DescribeLogGroups

* Define JSON policy field order that matches AWS examples

* Return the policy JSON string to the UI as string

This is consistent with how this was done before. The UI can then format and display this to the user on the Available Services page. This is the policy that the user will create the AWS user for the AWS integration with.

* Specify the correct AWS policy version

* Change AWS Policy to a list

* Add missing permissions

Based on this [sample KCL code](https://github.com/aws/aws-sdk-java-v2/issues/1214#issuecomment-488882831)

* Update version to 3.1.0-beta.2-SNAPSHOT

* Remove uneeded import

* Add Available Services API response error documentation

* Throw InternalServerErrorException instead of JSON exception

This is a bit cleaner from the API consumption side. A nice short error message is now returned instead of an obscure JSON error.