Add a test script to send Syslog TCP messages to Graylog.
In order to test Syslog-based plugins, we need the ability to generate traffic. For plugins like Palo Alto, where the traffic comes from a proprietary device, that can be quite difficult.  This script will send Syslog TCP messages with arbitrary payloads to a specified server/port, which allows us to take the payloads of real Palo logs and replay them as if they were coming from an actual Palo Alto device.

https://github.com/Graylog2/graylog-plugin-integrations/issues/483