Support AWS custom Input UI in Forwarder (#736)
* Re-enable AWS input for cloud

* Clean up URLs

- Clean up `Routes` object, since only one route in that object was
  being mounted and used
- Remove unused props from the previous cleanup

* Update Icon name

* Fix issues in components

- Move state update in `CloudWatch` inside `useEffect`
- Fix linter errors in `CloudWatch`
- Move styled component outside of render in `AWSAuthenticationTypes`,
  since that caused a console error

* Provide an embeddable version of CloudWatchApp

So far the custom Input UI was meant to be run in a full page on its
own, but in Forwarders we will embed that UI into another page. The
biggest change is removing page headers, but we may also do some styling
or other modifications eventually.

* Set up test environment

* Add option to submit Input externally

This converts the form data into a generic InputCreate request that can
be submitted to other resources, like the regular Inputs API or the
Forwarder API.

* Use formDataAdapter to AWS requests from form data

* Add frontend test to maven build

* Remove flow check from build

We don't use flow annotations on this plugin.

* Move form data into a fixture

Making tests easier to read and avoiding so much repetition.